refactor(map): convert terrain-resource part files to standalone classes (#3588)
Extract the `_TileMapGenTerrainResource` `part of 'tile_map_generator.dart'`
fragment (and its hardwood/noise extension part files) into standalone,
independently importable classes, eliminating the last part coupling in the
generation layer (only the whitelisted `tile_map_generator_types.dart` retains
`part of`).

- `TileMapGenTerrainResource`: public class (no longer a `part`), constructor-
  injected `TileMapGridGraph` plus strategy services.
- Strategy services split into focused standalone files, each under the 500
  non-comment-line target: `MountainRidgePlacer`, `TerrainRegionGrower`,
  `TerrainPatternRefiner`, `TerrainNoisePerturbation`, `HardwoodForestClusterer`.
- `terrain_blob_ops.dart`: shared component/interior/weighted-pick helpers,
  deduplicated across region-growing, refinement, noise, and hardwood passes.
- Protocol checker binding + `repo-lint.md` row refreshed to the public name
  (now 6 generator families implement `MapGenPass`).

Pure relocation; iteration order and RNG sequence unchanged (determinism
preserved). Added direct unit tests for the new standalone passes.

Refs #3588
