+++ SPEC/program/repo-lint.md
@@ -69,7 +69,7 @@
 | `tool/check_combat_no_raw_copies_in_resolution.dart` | Forbid raw ship-list clones `List.from(...)` / `List<ShipInstance>.from(...)` in `packages/colonizethis_combat/lib/src/combat/naval_combat_resolver.dart`; naval ship-list copies must delegate to the canonical `copyNavalShips(...)` helper. `List<Fleet>.from(...)` and other non-ship clones are out of scope; comment lines are ignored (Refs #3448); rule `repo.combat_no_raw_copies_in_resolution` |
 | `tool/check_map_grid_ops_central.dart` | Keep tile-grid deep-copy logic centralized in `packages/colonizethis_map/lib/**`: forbid any reference to the removed `copyTileMapGrid(` helper and any inline row-major deep copy `<g>.map((row) => row.toList()).toList()` outside the canonical helper file `packages/colonizethis_map/lib/src/tile_map_grid.dart`. Every defensive grid copy in the land-seed, lakes-province, and join-sea generation passes must route through `TileMapGrid.copy(...)` so the copy site has one source of truth; the canonical file is exempt and comment lines are ignored (Refs #3459); rule `repo.map_grid_ops_central` |
 | `tool/check_map_grid_cell_iteration_central.dart` | Keep tile-grid cell traversal centralized in `packages/colonizethis_map/lib/**`: forbid hand-rolled nested `for (var y …) { for (var x …) }` tile walks (a `for (<decl> y …)` header immediately followed by a `for (<decl> x …)` header) so iteration order — which seeded generation depends on for bit-for-bit determinism — has one definition in `TileMapGrid.forEachIndex(...)` / `TileMapGrid.forEachCell(...)`. Whole-file exemptions cover the canonical API home `src/tile_map_grid.dart`, the graph algorithms `src/gen/tile_map_grid_graph.dart`, and the two-pass distance transform `src/gen/tile_map_manhattan_distance_transform.dart`. A genuine per-loop exception (e.g. a `sync*` generator whose body must `yield`, or a bordered edge-skipping walk) carries an inline `ct-lint-allow: nested-grid-walk` marker on the outer loop header line or the comment line directly above it; comment lines are otherwise ignored (Refs #3574); rule `repo.map_grid_cell_iteration_central` |
-| `tool/check_map_gen_stage_protocol.dart` | Require the shared `MapGenStage` contract in `packages/colonizethis_map/lib/src/gen/map_gen_stage.dart` and `implements MapGenStage` (or the uniform `MapGenPass`) on each bound generator service family (`TileMapGenLandSeeds`, `TileMapGenLakesProvinces`, `ContinentJoinPass`, `TerrainJitterPass`, `SeaZoneSubdividePass`, `_TileMapGenTerrainResource`) so pass orchestration shares a uniform params + grid-in/grid-out protocol; the former join-sea family was split into three standalone `MapGenPass` passes (Refs #3459, #3574, #3588); rule `repo.map_gen_stage_protocol` |
+| `tool/check_map_gen_stage_protocol.dart` | Require the shared `MapGenStage` contract in `packages/colonizethis_map/lib/src/gen/map_gen_stage.dart` and `implements MapGenStage` (or the uniform `MapGenPass`) on each bound generator service family (`TileMapGenLandSeeds`, `TileMapGenLakesProvinces`, `ContinentJoinPass`, `TerrainJitterPass`, `SeaZoneSubdividePass`, `TileMapGenTerrainResource`) so pass orchestration shares a uniform params + grid-in/grid-out protocol; the former join-sea family was split into three standalone `MapGenPass` passes (Refs #3459, #3574, #3588); rule `repo.map_gen_stage_protocol` |
 | `tool/check_map_public_barrel_surface.dart` | Keep the public `colonizethis_map` barrel `packages/colonizethis_map/lib/colonizethis_map.dart` narrowed to intentional entry points: forbid any `export` of the internal-only generation primitives `src/gen/grid_voronoi.dart` (deterministicNoise / assignCellsToNearestSeed), `src/gen/topology_inference.dart` (inferTopologyFromTileMap), and `src/gen/tile_map_grid_graph.dart` (TileMapGridGraph). Same-package tests import these from `src/` directly; comment lines and `import` directives are ignored (Refs #3459 AC4); rule `repo.map_public_barrel_surface` |
 | `tool/check_map_region_data_access_central.dart` | Keep old/new-world region data selection centralized in `packages/colonizethis_map/lib/**`: forbid any inline `.oldWorld.provinces`, `.oldWorld.units`, `.newWorld.provinces`, or `.newWorld.units` region branch outside the canonical helper file `packages/colonizethis_map/lib/src/region_data_access.dart`. Every region-scoped provinces/units read in the view-building (orchestration, cell/unit) and ownership-overlay passes must resolve the region via `regionDataForMapRegionId(world, regionId)` so the old/new selection has one source of truth; the canonical file is exempt and comment lines are ignored (Refs #3459 AC3); rule `repo.map_region_data_access_central` |
 | `tool/check_map_region_dispatch_central.dart` | Keep old/new-world region *dispatch* centralized in `packages/colonizethis_map/lib/**`: forbid any inline `==` / `!=` comparison against `kRegionOldWorld` or `kRegionNewWorld` outside the canonical dispatch file `packages/colonizethis_map/lib/src/map_region_dispatch.dart`. Colour, capital-scope, port-tile bucketing, and view-data assembly paths must resolve old/new branching via `selectByMapRegionId(...)` / `selectByMapRegionIdOrNull(...)` so dispatch has one source of truth. Using the constants as map keys, list elements, named arguments, or string-interpolation operands is unaffected; the canonical file is exempt and comment lines are ignored (Refs #3574); rule `repo.map_region_dispatch_central` |
